Integrate market manager with core systems (#411)
