+++ Intersect.Server.Core/Database/PlayerData/Market/MarketStatisticsManager.cs
@@ -1,12 +1,13 @@
+using System;
 using System.Collections.Generic;
 
 namespace Intersect.Server.Database.PlayerData.Market;
 
-public class MarketStatisticsManager
+public static class MarketStatisticsManager
 {
-    private readonly Dictionary<int, MarketStatistics> _statistics = new();
+    private static readonly Dictionary<Guid, MarketStatistics> _statistics = new();
 
-    private MarketStatistics GetOrCreateStats(int itemId)
+    private static MarketStatistics GetOrCreateStats(Guid itemId)
     {
         if (!_statistics.TryGetValue(itemId, out var stats))
         {
@@ -17,14 +18,19 @@ public class MarketStatisticsManager
         return stats;
     }
 
-    public void RecordListing(int itemId, long price) => GetOrCreateStats(itemId).Record(price);
+    public static void RecordListing(Guid itemId, long price) => GetOrCreateStats(itemId).Record(price);
 
-    public void RecordSale(int itemId, long price) => GetOrCreateStats(itemId).Record(price);
+    public static void RecordSale(Guid itemId, long price) => GetOrCreateStats(itemId).Record(price);
 
-    public (long suggested, long min, long max) GetStatistics(int itemId)
+    public static (long suggested, long min, long max) GetStatistics(Guid itemId)
     {
         var stats = GetOrCreateStats(itemId);
         return (stats.SuggestedPrice, stats.MinPrice, stats.MaxPrice);
+    }
+
+    public static void UpdateStatistics(MarketTransaction transaction)
+    {
+        RecordSale(transaction.ItemId, transaction.Price);
     }
 }
